+++ app/components/Chakra.js
@@ -1,4 +1,5 @@
 "use client";
+
 // e.g. src/Chakra.js
 // a) import `ChakraProvider` component as well as the storageManagers
 import {
