+++ app/components/Modal/Auth/Oauth/OAuthButtons.jsx
@@ -13,11 +13,7 @@ import {
   TwitterIcon,
   YahooIcon,
 } from "../../../Icons/Components/IconComponents";
-import {
-  discordButtonLoading,
-  socialOnboardingAtom,
-} from "../../../atoms/authAtom";
-import { twitchButtonLoading } from "../../../atoms/authAtom";
+import { socialOnboardingAtom } from "../../../atoms/authAtom";
 import Cookies from "js-cookie";
 import {
   SocialOnboardingModalAtom,
@@ -26,14 +22,10 @@ import {
 import { useState } from "react";
 
 function OAuthButtons() {
-  const toast = useToast();
-
+  // States
   const [googleLoading, setGoogleLoading] = useState(false);
   const [socialOnboarding, setSocialOnboarding] =
     useRecoilState(socialOnboardingAtom);
-  const [discordLoading, setDiscordLoading] =
-    useRecoilState(discordButtonLoading);
-  const [twitchLoading, setTwitchLoading] = useRecoilState(twitchButtonLoading);
   const [authModalState, setAuthModalState] = useRecoilState(authModalAtom);
   const [socialOnboardingModal, setSocialOnboardingModal] = useRecoilState(
     SocialOnboardingModalAtom
@@ -111,7 +103,6 @@ function OAuthButtons() {
   const handleDiscordSignin = async () => {
     try {
       console.log("Handle discord sign in started.");
-      setDiscordLoading(true);
       const discordClientId = process.env.NEXT_PUBLIC_DISCORD_CLIENT_ID;
       const discordAuthUrl = `https://discord.com/api/oauth2/authorize?client_id=${discordClientId}&redirect_uri=${window.location.origin}/auth/discord&response_type=code&scope=identify%20email`;
       const width = 600; // Popup pencerenin genişliği
@@ -125,7 +116,6 @@ function OAuthButtons() {
       const interval = window.setInterval(() => {
         if (popupWindow.closed) {
           console.log("Discord popup closed.");
-          setDiscordLoading(false);
           // Clear the interval once the popup is closed
           window.clearInterval(interval);
         }
@@ -137,7 +127,6 @@ function OAuthButtons() {
   const handleTwitchSignin = async () => {
     try {
       console.log("Handle twitch sign in started.");
-      setTwitchLoading(true);
       const twitchClientId = process.env.NEXT_PUBLIC_TWITCH_CLIENT_ID;
       const twitchAuthUrl = `https://id.twitch.tv/oauth2/authorize?response_type=code&client_id=${twitchClientId}&redirect_uri=${window.location.origin}/auth/twitch&scope=user%3Aread%3Aemail&force_verify=true`;
       const width = 600; // Popup pencerenin genişliği
@@ -151,7 +140,6 @@ function OAuthButtons() {
       const interval = window.setInterval(() => {
         if (popupWindow.closed) {
           console.log("Twitch popup closed.");
-          setTwitchLoading(false);
           // Clear the interval once the popup is closed
           window.clearInterval(interval);
         }
@@ -215,7 +203,7 @@ function OAuthButtons() {
         {" • "}
         <Box>
           <IconButton
-            isLoading={discordLoading}
+            /* isLoading={discordLoading} */
             onClick={handleDiscordSignin}
             aria-label="Sign in with Discord"
             icon={<DiscordIcon h="10" w="10" color="#5562ea" />}
@@ -259,7 +247,7 @@ function OAuthButtons() {
         {" • "}
         <Box>
           <IconButton
-            isLoading={twitchLoading}
+            /* isLoading={twitchLoading} */
             onClick={handleTwitchSignin}
             aria-label="Sign in with Twitch"
             icon={<CustomAnimatedTwitchIcon w="10" h="10" fill="#a970ff" />}
