+++ app/components/Profile/Tabs/UserCommunities.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useRecoilState } from "recoil";
-import { profileInfoAtom } from "../atoms/postsAtom";
+import { profileInfoAtom } from "../../atoms/postsAtom";
 import { Box, Text } from "@chakra-ui/react";
 
 const UserCommunities = () => {
