+++ app/components/Profile/Tabs/UserFollowers.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { useRecoilState } from "recoil";
-import { profileInfoAtom } from "../atoms/postsAtom";
+import { profileInfoAtom } from "../../atoms/postsAtom";
 import { Box, Text } from "@chakra-ui/react";
 
-const UserPosts = () => {
+const UserFollowers = () => {
   const [profileOwnerInfo, setProfileOwnerInfo] =
     useRecoilState(profileInfoAtom);
   return (
@@ -13,10 +13,10 @@ const UserPosts = () => {
           // @ts-ignore
           profileOwnerInfo.Display_Name
         }
-        &apos; posts
+        &apos; followers
       </Text>
     </Box>
   );
 };
 
-export default UserPosts;
+export default UserFollowers;
